# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Book Library API",
    "version": "1.0.0",
    "description": "API for managing a library of books",
    "contact": {
      "name": "Your Name",
      "email": "youremail@example.com"
    }
  },

  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get a list of books",
        "responses": {
          "200": {
            "description": "List of books retrieved successfully",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "title": "Sample Book 1",
                    "author": "Author 1",
                    "publishedYear": 2020
                  },
                  {
                    "id": 2,
                    "title": "Sample Book 2",
                    "author": "Author 2",
                    "publishedYear": 2019
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new book",
        "requestBody": {
          "description": "New book data",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "New Book",
                "author": "New Author",
                "publishedYear": 2023
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book added successfully"
          }
        }
      }
    },
    "/books/{bookId}": {
      "get": {
        "summary": "Get book details by ID",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book details retrieved successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "title": "Sample Book 1",
                  "author": "Author 1",
                  "publishedYear": 2020
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "put": {
        "summary": "Update book details by ID",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Updated book data",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Updated Book",
                "author": "Updated Author",
                "publishedYear": 2024
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book details updated successfully"
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "delete": {
        "summary": "Delete book by ID",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    }
  }
}